Move Merge implementation into respective types for implementation
Tested-by: Kristofer Rye <kristofer.rye@gmail.com>
