+++ src/merge.rs
@@ -1,0 +1,5 @@
+pub trait Merge {
+	type Output;
+
+	fn merge(&self, other: &Self) -> Self::Output;
+}
